Fix route mapping.
Use CIDR notation when defining and checking routes to avoid confusion.

There was a problem when checking if a route was already mapped. For example, if the IP 198.162.0.1 already has a route defined and the program checks for the IP 198.162.0.11, the regex would match against the one ending with .1, and the .11 would not get mapped.